+++ src/onboarding/types.ts
@@ -1,0 +1,6 @@
+export enum AdventureCardName {
+  Add = 'add',
+  Dapp = 'dapp',
+  Profile = 'profile',
+  Learn = 'learn',
+}
